added custom exception handler & fixed userDTO for better response/request

+++ src/main/java/com/example/lab1/service/GenreService.java
@@ -2,6 +2,8 @@ package com.example.lab1.service;
 
 import com.example.lab1.model.Genre;
 import com.example.lab1.repository.GenreRepository;
+import com.example.lab1.service.customException.DuplicateResourceException;
+import com.example.lab1.service.customException.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,19 +22,24 @@ public class GenreService {
         return genreRepository.findAll();
     }
 
-    public Genre findByName(String name) {
+    public Genre findByName(String name) throws ResourceNotFoundException {
+        if(genreRepository.findByName(name) == null) {
+            throw new ResourceNotFoundException("Genre with name " + name + " not found");
+        }
         return genreRepository.findByName(name);
     }
 
-    public Genre save(Genre genre) {
+    public Genre save(Genre genre) throws DuplicateResourceException {
+        if(genreRepository.findByName(genre.getName()) != null) {
+            throw new DuplicateResourceException("Genre with name " + genre.getName() + " already exists");
+        }
         return genreRepository.save(genre);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws ResourceNotFoundException {
+        if(!genreRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Genre with id " + id + " not found");
+        }
         genreRepository.deleteById(id);
-    }
-
-    public Genre getGenreByName(String name) {
-        return genreRepository.findByName(name);
     }
 }
