+++ src/main/java/com/example/lab1/service/UserService.java
@@ -2,6 +2,8 @@ package com.example.lab1.service;
 
 import com.example.lab1.model.User;
 import com.example.lab1.repository.UserRepository;
+import com.example.lab1.service.customException.DuplicateResourceException;
+import com.example.lab1.service.customException.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,19 +21,26 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public Optional<User> findByUsername(String username) {
+    public Optional<User> findByUsername(String username) throws ResourceNotFoundException {
+
+        Optional<User> user = userRepository.findByUsername(username);
+        if (user.isEmpty()) {
+            throw new ResourceNotFoundException("User with username " + username + " not found");
+        }
         return userRepository.findByUsername(username);
     }
 
-    public Optional<User> findByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    public User save(User user) {
+    public User save(User user) throws DuplicateResourceException {
+        if (userRepository.findByUsername(user.getUsername()).isPresent()) {
+            throw new DuplicateResourceException("User with username " + user.getUsername() + " already exists");
+        }
         return userRepository.save(user);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws ResourceNotFoundException {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new ResourceNotFoundException("User with id " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
 }
